Add the flag/switching mechanism to allow us to render new details pages
By doing the flagging we can check the pages in slowly without causing disruption, it will also allow us to roll things back if a bug is found last minute. By default we're still showing legacy pages but we will flip that once pages are added.
