+++ shell/pages/explorer/resource/detail/configmap.vue
@@ -1,0 +1,10 @@
+<script setup lang="ts">
+import DetailPage from '@shell/components/Resource/Detail/Page.vue';
+</script>
+<template>
+  <DetailPage>
+    <template #top-area>
+      ConfigMap
+    </template>
+  </DetailPage>
+</template>
